+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Gemini models are available (gemini-2.0-flash and gemini-2.0-flash-lite-preview)
 - Generated files are displayed under the underneath the response
 - Costs for (some) individual functionalities are displayed